Cloud Computing/fastapi-serving/requirements.txt
fastapi==0.95.0
numpy==1.23.5
pydantic==1.10.7
tensorflow==2.12.0
uvicorn==0.21.1
Pillow==9.5.0